refactor: remove imports in original dto
