+++ tests/libecalc/core/models/compressor_modelling/test_compressor_with_turbine.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from libecalc import dto
+from libecalc.domain.process import dto
 from libecalc.domain.process.core.compressor.base import CompressorWithTurbineModel
