+++ tests/libecalc/core/models/compressor_modelling/test_variable_speed_compressor_train_multiple_streams.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 import libecalc.common.fixed_speed_pressure_control
-from libecalc import dto
+from libecalc.domain.process import dto
 from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureControl
 from libecalc.common.fluid import MultipleStreamsAndPressureStream
 from libecalc.common.fluid_stream_type import FluidStreamType
@@ -20,7 +20,7 @@ from libecalc.domain.process.core.compressor.train.variable_speed_compressor_tra
 from libecalc.domain.process.core.results.compressor import (
     CompressorTrainCommonShaftFailureStatus,
 )
-from libecalc.dto import InterstagePressureControl
+from libecalc.domain.process.dto import InterstagePressureControl
 
 
 def calculate_relative_difference(value1, value2):
