+++ tests/libecalc/core/models/test_pump.py
@@ -4,7 +4,7 @@ import pytest
 
 from libecalc.common.serializable_chart import ChartCurveDTO, SingleSpeedChartDTO, VariableSpeedChartDTO
 from libecalc.domain.process.core.chart import SingleSpeedChart, VariableSpeedChart
-from libecalc.domain.process.pump import PumpSingleSpeed, PumpVariableSpeed
+from libecalc.domain.process.pump.pump import PumpSingleSpeed, PumpVariableSpeed
 from libecalc.domain.process.pump.pump import _adjust_for_head_margin
